services/collector/requirements.txt
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
